frontend/src/app/components/form/registrazione-cliente/registrazione-cliente.component.ts
import { Component, EventEmitter, Output } from '@angular/core';
import { MatButtonModule } from '@angular/material/button';
import { MatInputModule } from '@angular/material/input';
import { MatFormFieldModule } from '@angular/material/form-field';
import { MatIconModule } from '@angular/material/icon';
import {
  FormControl,
  FormGroup,
  ReactiveFormsModule,
  Validators,
} from '@angular/forms';
import { Router } from '@angular/router';

@Component({
  selector: 'app-registrazione-cliente',
  standalone: true,
  imports: [
    MatFormFieldModule,
    MatInputModule,
    MatButtonModule,
    MatIconModule,
    ReactiveFormsModule,
  ],
  templateUrl: './registrazione-cliente.component.html',
  styleUrl: './registrazione-cliente.component.css',
})
export class RegistrazioneClienteFormComponent {
  hidePassword = true;

  form = new FormGroup({
    name: new FormControl('', Validators.required),
    surname: new FormControl('', Validators.required),
    email: new FormControl('', [
      Validators.required,
      Validators.pattern(/^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$/),
    ]),
    password: new FormControl('', Validators.required),
  });

  constructor(router: Router) {}

  @Output() onSubmit = new EventEmitter<FormGroup>();

  on_submit() {
    if (this.form.valid) {
      this.onSubmit.emit(this.form);
    }
  }
}
